+++ src/main/java/com/example/rolebase/controller/AdminController.java
@@ -32,7 +32,7 @@ public class AdminController implements AdminApi {
 
     @Override
     @DeleteMapping("/delete-user/{id}")
-    public ResponseEntity<String> deleteUser(@PathVariable Integer id) {
+    public ResponseEntity<String> deleteUser(@PathVariable Long id) {
         userService.deleteUser(id);
         return ResponseEntity.ok("User with ID " + id + " deleted successfully.");
     }
